moving new index APIs to _ex files
