+++ c_api/IndexIVF_c_ex.cpp
@@ -1,0 +1,22 @@
+/**
+ * Copyright (c) Facebook, Inc. and its affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+// Copyright 2023-present Facebook. All Rights Reserved.
+// -*- c++ -*-
+#include "IndexIVF_c.h"
+#include <faiss/IndexIVF.h>
+#include "macros_impl.h"
+
+using faiss::IndexIVF;
+
+int faiss_IndexIVF_set_direct_map(FaissIndexIVF* index, int direct_map_type) {
+    try {
+        reinterpret_cast<IndexIVF*>(index)->set_direct_map_type(
+                static_cast<faiss::DirectMap::Type>(direct_map_type));
+    }
+    CATCH_AND_HANDLE
+}
