+++ c_api/Index_c.cpp
@@ -157,22 +157,6 @@ int faiss_Index_reconstruct(const FaissIndex* index, idx_t key, float* recons) {
     CATCH_AND_HANDLE
 }
 
-int faiss_Index_reconstruct_batch(const FaissIndex* index, idx_t n, const idx_t* keys, float* recons) {
-    try {
-        reinterpret_cast<const faiss::Index*>(index)->reconstruct_batch(n, keys, recons);
-    }
-    CATCH_AND_HANDLE
-}
-
-
-int faiss_Index_merge_from(FaissIndex* index, FaissIndex* other, const idx_t add_id) {
-    try {
-        reinterpret_cast<faiss::Index*>(index)->merge_from(
-                *reinterpret_cast<faiss::Index*>(other), add_id);
-    }
-    CATCH_AND_HANDLE
-}
-
 int faiss_Index_reconstruct_n(
         const FaissIndex* index,
         idx_t i0,
